Fix data

+++ bedwars-api/src/main/java/com/andrei1058/bedwars/api/server/VersionSupport.java
@@ -262,7 +262,7 @@ public abstract class VersionSupport {
      */
     public abstract ItemStack colourItem(ItemStack itemStack, ITeam bedWarsTeam);
 
-    public abstract ItemStack createItemStack(String material, int amount, short data);
+    public abstract ItemStack createItemStack(String material, int amount, byte data);
 
     public abstract void teamCollideRule(Team team);
